@@ -2,12 +2,69 @@ package dictionarygen
 
 import (
 	"io"
+	"strconv"
 
-	"layeh.com/radius/dictionary"
+	"github.com/ParspooyeshFanavar/go-radius/dictionary"
 )
 
 func (g *Generator) genVendor(w io.Writer, vendor *dictionary.Vendor) {
 	ident := identifier(vendor.Name)
+
+	//attrOIDMap
+	p(w)
+	p(w, `var attrOIDMap = map[radius.Type]radius.NameType {`)
+	for _, attr := range vendor.Attributes {
+		valueMapping := "nil"
+		for _, value := range vendor.Values {
+			if value.Attribute == attr.Name {
+				valueMapping = identifier(value.Attribute) + `_GetValueString`
+			}
+		}
+
+		p(w, `	`, strconv.Itoa(attr.OID[0]), `:{ "`+attr.Name+`", `, strconv.Itoa(int(attr.Type)), `, `, valueMapping, `},`)
+	}
+	p(w, `}`)
+
+	//attrNameMap
+	p(w)
+	p(w, `var attrNameMap = map[string]radius.OIDType {`)
+	for _, attr := range vendor.Attributes {
+		valueMapping := "nil"
+		for _, value := range vendor.Values {
+			if value.Attribute == attr.Name {
+				valueMapping = identifier(value.Attribute) + `_GetValueNumber`
+			}
+		}
+
+		p(w, `	`, `"`, attr.Name, `"`, `:{ `+strconv.Itoa(attr.OID[0])+`, `, strconv.Itoa(int(attr.Type)), `, `, valueMapping, `},`)
+	}
+	p(w, `}`)
+
+	//GetAttrName function
+	p(w)
+	p(w, `func GetAttrName(T byte) (string, dictionary.AttributeType, func(uint32) (string, error)) {`)
+	p(w, `	`, `name, ok := attrOIDMap[radius.Type(T)]`)
+	p(w, `	`, `if ok {`)
+	p(w, `	`, `	`, `return name.Name, name.T, name.ValueMapFunc`)
+	p(w, `	`, `}`)
+	p(w, `	`, `return "", 2, nil`)
+	p(w, `}`)
+
+	//GetAttrOID function
+	p(w)
+	p(w, `func GetAttrOID(name string) (radius.Type, dictionary.AttributeType, func(string) (uint32, error)) {`)
+	p(w, `	`, `t, ok := attrNameMap[name]`)
+	p(w, `	`, `if ok {`)
+	p(w, `	`, `	`, `return t.OID, t.T, t.ValueMapFunc`)
+	p(w, `	`, `}`)
+	p(w, `	`, `return -1, dictionary.AttributeOctets, nil`)
+	p(w, `}`)
+
+	//Register Vendor attributeMap
+	p(w)
+	p(w, `func init() {`)
+	p(w, `	`, `attributemap.RegisterVendor(`, `_`, ident, `_VendorID`, `, GetAttrName, GetAttrOID)`)
+	p(w, `}`)
 
 	p(w)
 	p(w, `func _`, ident, `_AddVendor(p *radius.Packet, typ byte, attr radius.Attribute) (err error) {`)
