@@ -1,3 +1,3 @@
-//go:generate go run ../cmd/radius-dict-gen/main.go -package rfc2867 -output generated.go -ref Acct-Status-Type:layeh.com/radius/rfc2866 /usr/share/freeradius/dictionary.rfc2867
+//go:generate go run ../cmd/radius-dict-gen/main.go -package rfc2867 -output generated.go -ref Acct-Status-Type:github.com/ParspooyeshFanavar/go-radius/rfc2866 /usr/share/freeradius/dictionary.rfc2867
 
 package rfc2867
