@@ -1,3 +1,3 @@
-//go:generate go run ../cmd/radius-dict-gen/main.go -package rfc3576 -output generated.go -ref Service-Type:layeh.com/radius/rfc2865 /usr/share/freeradius/dictionary.rfc3576
+//go:generate go run ../cmd/radius-dict-gen/main.go -package rfc3576 -output generated.go -ref Service-Type:github.com/ParspooyeshFanavar/go-radius/rfc2865 /usr/share/freeradius/dictionary.rfc3576
 
 package rfc3576
