@@ -1,3 +1,3 @@
-//go:generate go run ../cmd/radius-dict-gen/main.go -package rfc5176 -output generated.go -ref Error-Cause:layeh.com/radius/rfc3576 /usr/share/freeradius/dictionary.rfc5176
+//go:generate go run ../cmd/radius-dict-gen/main.go -package rfc5176 -output generated.go -ref Error-Cause:github.com/ParspooyeshFanavar/go-radius/rfc3576 /usr/share/freeradius/dictionary.rfc5176
 
 package rfc5176
